+++ packages/flutter_solo/lib/flutter_solo.dart
@@ -11,6 +11,8 @@ library;
 
 export 'package:flutter/foundation.dart' show ValueListenable;
 export 'package:solo/solo.dart';
+export 'src/solo_builder.dart';
 export 'src/solo_listenable.dart';
+export 'src/solo_select_builder.dart';
 export 'src/solo_selection.dart' show SoloSelection;
 export 'src/solo_selector.dart';